wip: add RANDOM ROWID table option
With the new RANDOM ROWID keywords, a table can explicitly
state that it wants its rowid to be generated randomly,
not consecutively, without having to previously create a sentinel
record with rowid == max(i64).
RANDOM ROWID is mutually exclusive with WITHOUT ROWID,
and cannot be combined with AUTOINCREMENT primary key.

Fixes #12