clarify README
From Abdur's feedback, it is important to justify the fork right in the
README instead of just pointing to the motivation page. So bring some of
the text - with adaptations for the README format - from the motivation
page back.

We have also heard feedback that the file compatibility claim is not
clear enough regarding future extensions like CRC that may require file
modifications, so clarify that as well.