sqld,wal_hook: support nesting bottomless WAL in wal_hook
Both wal_hook and bottomless use the same nifty trick of allocating
an additional field in WAL methods struct and using it to store a pointer
to the original methods. This trick stops working if two or more impls
try to use the same offset though, so wal_hook now takes into account that
an underlying impl can also use an extra field.