libsql/core: Implement Drop for Statement
This implements Drop trait for the Statement struct to clean up after
ourselves once we no longer need the statement. One complication is that
Rows also holds a reference to the sqlite3_stmt object, which is why I
introduced a StatementInner. We can probably make this cleaner, but
that's for later.