libsql/core: Add `v2` api
This PR adds a new in-progress v2 rewrite of the libsql API to support
both a sqlite3 and http based backend. None of this code affects the
current API and is written in such a way to migrate over easily.

The current approach to abstract both the sqlite3 and http api is to use
internal trait objects so that the public types do not contain generics.
In the future, if we find that trait objects are too slow we can move to
enums.