crates/replication: do not wait for replication in execute()
Waiting makes sense if we have a background fiber that continuously
replicates frames back to us. In libsql crate we don't do that,
so we effectively wait for frames to be replicated, but no replication
was ever requested, and we deadlock.
Instead, as you can see in the modified examples/remote_sync example,
users call db.sync().await() in order to be able to read their own
writes.