libsql/replication: Write delegation v2 with txn
This implements write delegation in the proper way by not leveraging the
internal libsql parser but to use a separate rust one. This allows us
to not corrupt libsql connection state when we want to submit remote
executes.

What this means is there is now a `RemoteConnection` implementation that
wraps the libsql version. It will detect when sql statements are write
based and will submit those against the primary.