Add more logging in case snapshots fail (#1270)
We have a user with a failing snapshot, but we don't know why.
We know rename returns EINVAL, which could happen for a variety
of reasons, all of them related to the paths themselves.

By printing the path, we will be able to figure out the actual reason.
Right now we're shooting in the dark.